Refactor PolarisMetaStoreManager interface into multiple interfaces
