+++ polaris-core/src/main/java/org/apache/polaris/core/persistence/LocalPolarisMetaStoreManagerFactory.java
@@ -25,6 +25,7 @@ import java.util.function.Supplier;
 import org.apache.polaris.core.PolarisCallContext;
 import org.apache.polaris.core.PolarisDefaultDiagServiceImpl;
 import org.apache.polaris.core.PolarisDiagnostics;
+import org.apache.polaris.core.auth.PolarisSecretsManager.PrincipalSecretsResult;
 import org.apache.polaris.core.context.CallContext;
 import org.apache.polaris.core.context.RealmContext;
 import org.apache.polaris.core.entity.PolarisEntity;
@@ -75,15 +76,14 @@ public abstract class LocalPolarisMetaStoreManagerFactory<StoreType>
   }
 
   @Override
-  public synchronized Map<String, PolarisMetaStoreManager.PrincipalSecretsResult> bootstrapRealms(
-      List<String> realms) {
-    Map<String, PolarisMetaStoreManager.PrincipalSecretsResult> results = new HashMap<>();
+  public synchronized Map<String, PrincipalSecretsResult> bootstrapRealms(List<String> realms) {
+    Map<String, PrincipalSecretsResult> results = new HashMap<>();
 
     for (String realm : realms) {
       RealmContext realmContext = () -> realm;
       if (!metaStoreManagerMap.containsKey(realmContext.getRealmIdentifier())) {
         initializeForRealm(realmContext);
-        PolarisMetaStoreManager.PrincipalSecretsResult secretsResult =
+        PrincipalSecretsResult secretsResult =
             bootstrapServiceAndCreatePolarisPrincipalForRealm(
                 realmContext, metaStoreManagerMap.get(realmContext.getRealmIdentifier()));
         results.put(realmContext.getRealmIdentifier(), secretsResult);
@@ -157,9 +157,8 @@ public abstract class LocalPolarisMetaStoreManagerFactory<StoreType>
    * metastore and creates a root service principal. After that we rotate the root principal
    * credentials and print them to stdout
    */
-  private PolarisMetaStoreManager.PrincipalSecretsResult
-      bootstrapServiceAndCreatePolarisPrincipalForRealm(
-          RealmContext realmContext, PolarisMetaStoreManager metaStoreManager) {
+  private PrincipalSecretsResult bootstrapServiceAndCreatePolarisPrincipalForRealm(
+      RealmContext realmContext, PolarisMetaStoreManager metaStoreManager) {
     // While bootstrapping we need to act as a fake privileged context since the real
     // CallContext hasn't even been resolved yet.
     PolarisCallContext polarisContext =
@@ -199,7 +198,7 @@ public abstract class LocalPolarisMetaStoreManagerFactory<StoreType>
                     .getInternalPropertiesAsMap()
                     .get(PolarisEntityConstants.getClientIdPropertyName()))
             .getPrincipalSecrets();
-    PolarisMetaStoreManager.PrincipalSecretsResult rotatedSecrets =
+    PrincipalSecretsResult rotatedSecrets =
         metaStoreManager.rotatePrincipalSecrets(
             polarisContext,
             secrets.getPrincipalClientId(),
