+++ polaris-core/src/main/java/org/apache/polaris/core/persistence/resolver/Resolver.java
@@ -36,11 +36,12 @@ import org.apache.polaris.core.entity.PolarisEntityId;
 import org.apache.polaris.core.entity.PolarisEntityType;
 import org.apache.polaris.core.entity.PolarisGrantRecord;
 import org.apache.polaris.core.entity.PolarisPrivilege;
-import org.apache.polaris.core.persistence.PolarisMetaStoreManager;
 import org.apache.polaris.core.persistence.cache.EntityCache;
 import org.apache.polaris.core.persistence.cache.EntityCacheByNameKey;
 import org.apache.polaris.core.persistence.cache.EntityCacheEntry;
 import org.apache.polaris.core.persistence.cache.EntityCacheLookupResult;
+import org.apache.polaris.core.persistence.cache.PolarisRemoteCache;
+import org.apache.polaris.core.persistence.cache.PolarisRemoteCache.ChangeTrackingResult;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -57,7 +58,7 @@ public class Resolver {
   private final @NotNull PolarisDiagnostics diagnostics;
 
   // the polaris metastore manager
-  private final @NotNull PolarisMetaStoreManager metaStoreManager;
+  private final @NotNull PolarisRemoteCache polarisRemoteCache;
 
   // the cache of entities
   private final @NotNull EntityCache cache;
@@ -110,7 +111,7 @@ public class Resolver {
    * Constructor, effectively starts an entity resolver session
    *
    * @param polarisCallContext the polaris call context
-   * @param metaStoreManager meta store manager
+   * @param polarisRemoteCache meta store manager
    * @param callerPrincipalId if not 0, the id of the principal calling the service
    * @param callerPrincipalName if callerPrincipalId is 0, the name of the principal calling the
    *     service
@@ -126,7 +127,7 @@ public class Resolver {
    */
   public Resolver(
       @NotNull PolarisCallContext polarisCallContext,
-      @NotNull PolarisMetaStoreManager metaStoreManager,
+      @NotNull PolarisRemoteCache polarisRemoteCache,
       long callerPrincipalId,
       @Nullable String callerPrincipalName,
       @Nullable Set<String> callerPrincipalRoleNamesScope,
@@ -134,7 +135,7 @@ public class Resolver {
       @Nullable String referenceCatalogName) {
     this.polarisCallContext = polarisCallContext;
     this.diagnostics = polarisCallContext.getDiagServices();
-    this.metaStoreManager = metaStoreManager;
+    this.polarisRemoteCache = polarisRemoteCache;
     this.cache = cache;
     this.callerPrincipalName = callerPrincipalName;
     this.callerPrincipalId = callerPrincipalId;
@@ -144,7 +145,7 @@ public class Resolver {
     this.callerPrincipalRoleNamesScope = callerPrincipalRoleNamesScope;
 
     // validate inputs
-    this.diagnostics.checkNotNull(metaStoreManager, "unexpected_null_metaStoreManager");
+    this.diagnostics.checkNotNull(polarisRemoteCache, "unexpected_null_polarisRemoteCache");
     this.diagnostics.checkNotNull(cache, "unexpected_null_cache");
     this.diagnostics.check(
         callerPrincipalId != 0 || callerPrincipalName != null, "principal_must_be_specified");
@@ -534,8 +535,8 @@ public class Resolver {
               .collect(Collectors.toList());
 
       // now get the current backend versions of all these entities
-      PolarisMetaStoreManager.ChangeTrackingResult changeTrackingResult =
-          this.metaStoreManager.loadEntitiesChangeTracking(this.polarisCallContext, entityIds);
+      ChangeTrackingResult changeTrackingResult =
+          this.polarisRemoteCache.loadEntitiesChangeTracking(this.polarisCallContext, entityIds);
 
       // refresh any entity which is not fresh. If an entity is missing, reload it
       Iterator<EntityCacheEntry> entityIterator = toValidate.iterator();
