+++ polaris-core/src/main/java/org/apache/polaris/core/storage/cache/StorageCredentialCache.java
@@ -30,7 +30,7 @@ import org.apache.iceberg.exceptions.UnprocessableEntityException;
 import org.apache.polaris.core.PolarisCallContext;
 import org.apache.polaris.core.entity.PolarisEntity;
 import org.apache.polaris.core.entity.PolarisEntityType;
-import org.apache.polaris.core.persistence.PolarisMetaStoreManager;
+import org.apache.polaris.core.storage.PolarisCredentialVendor;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.VisibleForTesting;
 import org.slf4j.Logger;
@@ -94,7 +94,7 @@ public class StorageCredentialCache {
   /**
    * Either get from the cache or generate a new entry for a scoped creds
    *
-   * @param metaStoreManager the meta storage manager used to generate a new scoped creds if needed
+   * @param credentialVendor the credential vendor used to generate a new scoped creds if needed
    * @param callCtx the call context
    * @param polarisEntity the polaris entity that is going to scoped creds
    * @param allowListOperation whether allow list action on the provided read and write locations
@@ -103,7 +103,7 @@ public class StorageCredentialCache {
    * @return the a map of string containing the scoped creds information
    */
   public Map<String, String> getOrGenerateSubScopeCreds(
-      @NotNull PolarisMetaStoreManager metaStoreManager,
+      @NotNull PolarisCredentialVendor credentialVendor,
       @NotNull PolarisCallContext callCtx,
       @NotNull PolarisEntity polarisEntity,
       boolean allowListOperation,
@@ -125,8 +125,8 @@ public class StorageCredentialCache {
     Function<StorageCredentialCacheKey, StorageCredentialCacheEntry> loader =
         k -> {
           LOGGER.atDebug().log("StorageCredentialCache::load");
-          PolarisMetaStoreManager.ScopedCredentialsResult scopedCredentialsResult =
-              metaStoreManager.getSubscopedCredsForEntity(
+          PolarisCredentialVendor.ScopedCredentialsResult scopedCredentialsResult =
+              credentialVendor.getSubscopedCredsForEntity(
                   k.getCallContext(),
                   k.getCatalogId(),
                   k.getEntityId(),
