+++ polaris-service/src/main/java/org/apache/polaris/service/BootstrapRealmsCommand.java
@@ -23,8 +23,8 @@ import io.dropwizard.core.setup.Bootstrap;
 import java.util.Map;
 import net.sourceforge.argparse4j.inf.Namespace;
 import org.apache.polaris.core.PolarisConfigurationStore;
+import org.apache.polaris.core.auth.PolarisSecretsManager.PrincipalSecretsResult;
 import org.apache.polaris.core.persistence.MetaStoreManagerFactory;
-import org.apache.polaris.core.persistence.PolarisMetaStoreManager;
 import org.apache.polaris.service.config.ConfigurationStoreAware;
 import org.apache.polaris.service.config.PolarisApplicationConfig;
 import org.apache.polaris.service.context.CallContextResolver;
@@ -61,13 +61,12 @@ public class BootstrapRealmsCommand extends ConfiguredCommand<PolarisApplication
     }
 
     // Execute the bootstrap
-    Map<String, PolarisMetaStoreManager.PrincipalSecretsResult> results =
+    Map<String, PrincipalSecretsResult> results =
         metaStoreManagerFactory.bootstrapRealms(configuration.getDefaultRealms());
 
     // Log any errors:
     boolean success = true;
-    for (Map.Entry<String, PolarisMetaStoreManager.PrincipalSecretsResult> result :
-        results.entrySet()) {
+    for (Map.Entry<String, PrincipalSecretsResult> result : results.entrySet()) {
       if (!result.getValue().isSuccess()) {
         LOGGER.error(
             "Bootstrapping `{}` failed: {}",
