+++ polaris-service/src/main/java/org/apache/polaris/service/admin/PolarisAdminService.java
@@ -57,6 +57,7 @@ import org.apache.polaris.core.admin.model.ViewPrivilege;
 import org.apache.polaris.core.auth.AuthenticatedPolarisPrincipal;
 import org.apache.polaris.core.auth.PolarisAuthorizableOperation;
 import org.apache.polaris.core.auth.PolarisAuthorizer;
+import org.apache.polaris.core.auth.PolarisGrantManager.LoadGrantsResult;
 import org.apache.polaris.core.catalog.PolarisCatalogHelpers;
 import org.apache.polaris.core.context.CallContext;
 import org.apache.polaris.core.entity.CatalogEntity;
@@ -1189,7 +1190,7 @@ public class PolarisAdminService {
     PolarisEntity principalEntity =
         findPrincipalByName(principalName)
             .orElseThrow(() -> new NotFoundException("Principal %s not found", principalName));
-    PolarisMetaStoreManager.LoadGrantsResult grantList =
+    LoadGrantsResult grantList =
         metaStoreManager.loadGrantsToGrantee(
             getCurrentPolarisContext(), principalEntity.getCatalogId(), principalEntity.getId());
     return buildEntitiesFromGrantResults(grantList, false, null);
@@ -1253,7 +1254,7 @@ public class PolarisAdminService {
         findPrincipalRoleByName(principalRoleName)
             .orElseThrow(
                 () -> new NotFoundException("PrincipalRole %s not found", principalRoleName));
-    PolarisMetaStoreManager.LoadGrantsResult grantList =
+    LoadGrantsResult grantList =
         metaStoreManager.loadGrantsOnSecurable(
             getCurrentPolarisContext(),
             principalRoleEntity.getCatalogId(),
@@ -1272,7 +1273,7 @@ public class PolarisAdminService {
    * @return list of grantees or securables matching the filter
    */
   private List<PolarisEntity> buildEntitiesFromGrantResults(
-      @NotNull PolarisMetaStoreManager.LoadGrantsResult grantList,
+      @NotNull LoadGrantsResult grantList,
       boolean grantees,
       @Nullable Function<PolarisGrantRecord, Boolean> grantFilter) {
     Map<Long, PolarisBaseEntity> granteeMap = grantList.getEntitiesAsMap();
@@ -1306,7 +1307,7 @@ public class PolarisAdminService {
         findPrincipalRoleByName(principalRoleName)
             .orElseThrow(
                 () -> new NotFoundException("PrincipalRole %s not found", principalRoleName));
-    PolarisMetaStoreManager.LoadGrantsResult grantList =
+    LoadGrantsResult grantList =
         metaStoreManager.loadGrantsToGrantee(
             getCurrentPolarisContext(),
             principalRoleEntity.getCatalogId(),
@@ -1532,7 +1533,7 @@ public class PolarisAdminService {
     PolarisEntity catalogRoleEntity =
         findCatalogRoleByName(catalogName, catalogRoleName)
             .orElseThrow(() -> new NotFoundException("CatalogRole %s not found", catalogRoleName));
-    PolarisMetaStoreManager.LoadGrantsResult grantList =
+    LoadGrantsResult grantList =
         metaStoreManager.loadGrantsOnSecurable(
             getCurrentPolarisContext(),
             catalogRoleEntity.getCatalogId(),
@@ -1551,7 +1552,7 @@ public class PolarisAdminService {
     PolarisEntity catalogRoleEntity =
         findCatalogRoleByName(catalogName, catalogRoleName)
             .orElseThrow(() -> new NotFoundException("CatalogRole %s not found", catalogRoleName));
-    PolarisMetaStoreManager.LoadGrantsResult grantList =
+    LoadGrantsResult grantList =
         metaStoreManager.loadGrantsToGrantee(
             getCurrentPolarisContext(),
             catalogRoleEntity.getCatalogId(),
