+++ polaris-service/src/main/java/org/apache/polaris/service/persistence/InMemoryPolarisMetaStoreManagerFactory.java
@@ -25,6 +25,7 @@ import java.util.Map;
 import java.util.Set;
 import java.util.function.Supplier;
 import org.apache.polaris.core.PolarisDiagnostics;
+import org.apache.polaris.core.auth.PolarisSecretsManager.PrincipalSecretsResult;
 import org.apache.polaris.core.context.RealmContext;
 import org.apache.polaris.core.persistence.LocalPolarisMetaStoreManagerFactory;
 import org.apache.polaris.core.persistence.PolarisMetaStoreManager;
@@ -70,11 +71,11 @@ public class InMemoryPolarisMetaStoreManagerFactory
   }
 
   private void bootstrapRealmAndPrintCredentials(String realmId) {
-    Map<String, PolarisMetaStoreManager.PrincipalSecretsResult> results =
+    Map<String, PrincipalSecretsResult> results =
         this.bootstrapRealms(Collections.singletonList(realmId));
     bootstrappedRealms.add(realmId);
 
-    PolarisMetaStoreManager.PrincipalSecretsResult principalSecrets = results.get(realmId);
+    PrincipalSecretsResult principalSecrets = results.get(realmId);
 
     String msg =
         String.format(
